// src/app/shared/shared/enum/context-menu/context-menu.enum.ts
export enum ContextMenu {
    Edit = 'Edit', // sửa
    Delete = 'Delete', // Xóa
    DisActive = 'DisActive', // ngừng kích hoạt tài khoản
    Active = 'Active', // kích hoạt tài khoản
    NotifyNewsfeed = 'NotifyNewsfeed', // Thông báo mạng xã hội nội bộ
    EmployeeSelfUpdate = 'EmployeeSelfUpdate', // Nhân viên tự cập nhật
    MergeData = 'MergeData', // Trộn văn bản
    Print = 'Print', // In
    Export = 'Export', // Xuất khẩu
    ExportExcel = 'ExportExcel', // Xuất khẩu Excel
    ExportPDF = 'ExportPDF', // Xuất khẩu PDF
    AdjustOrganization = 'AdjustOrganization', // Điều chuyển đơn vị công tác
    Import = 'Import', // Nhập khẩu
    UpdateFile = 'UpdateFile', // Cập nhật file
    UpdateFileMass = 'UpdateFileMass', // Cập nhật file hàng loạt
    AdvanceExport = 'AdvanceExport', // Cập nhật file hàng loạt
    ViewDocument = 'ViewDocument',
    Duplicate = 'Duplicate', // Nhân bản
    GenerateAppendix = 'GenerateAppendix', // sinh phụ lục hợp đồng
    UpdateOfficialDate = 'UpdateOfficialDate', // Cập nhật ngày chính thức
    UpdateInfo = 'UpdateInfo', // Cập nhật ngày chính thức
    TerminateContract = 'TerminateContract', // Chấm dứt hợp đồng
    CreateDocument = 'CreateDocument', // Tạo văn bản
    AppendixContractDocument = 'AppendixContractDocument', // Tạo văn bản
    DocumentContract = 'DocumentContract', // Tạo văn bản
    SendEmail = 'Gửi Email',
    SendSMS = 'Gửi SMS',
    SaveAndAdd = "SaveAndAdd",
    UpdateProfile = "UpdateProfile", // Cập nhật hồ sơ
    UpdateMultiAvatar = "UpdateProfile", // Cập nhật ảnh hàng loạt
    Employee = "Employee", //Nhập khẩu Hồ sơ
    EmployeeRelationship = "EmployeeRelationship", //Nhập khẩu Thông tin gia đình
    EmployeePromotion = "EmployeePromotion", //Nhập khẩu Quá trình công tác
    EmployeeEducation = "EmployeeEducation", //Nhập khẩu Bằng cấp
    EmployeeCertification = "EmployeeCertification", //Nhập khẩu Chứng chỉ
    EmployeeExperience = "EmployeeExperience", //Nhập khẩu Kinh nghiệm làm việc
    EmployeeSkill = "EmployeeSkill", //Nhập khẩu Kỹ năng
    EmployeeJobRole = "EmployeeJobRole", //Nhập khẩu Vai trò đảm nhiệm
    EmployeeDocument = "EmployeeDocument", //Nhập khẩu Vai trò đảm nhiệm
    ImportDocumentSeries = "ImportDocumentSeries", //Tải tài liệu hàng loạt
    AddProfile = "AddProfile", // Thêm hồ sơ
    ReceiveCandidatePass = "ReceiveCandidatePass", // Tiếp nhận ứng viên trúng tuyển
    IdentifiedEmployee = "IdentifiedEmployee", // Nhận diện hồ sơ tự động
    CustomColumn = "CustomColumn", // Tùy chỉnh cột
    RefreshData = "RefreshData",
    PrintExcel = "PrintExcel", //In dạng excel
    PrintPdf = "PrintPdf", //In dạng PDF
    ApproveUpdateProfile = "ApproveUpdateProfile", // Phê duyệt cập nhật hồ sơ
    RejectUpdateProfile = "RejectUpdateProfile", // Từ chối cập nhật hồ sơ
    SendRequestUpdateProfile = "SendRequestUpdateProfile"// Gửi yêu cầu cập nhật hồ sơ

}